Adding capability for maps to be run in realtime (#51)
* WIP.

* Combining drivers for skewt and maps.

* WIP

* This function doesn't belong here.

* get_transforms should handle item, list, and dict.

* Fix calls to get_transform

* Fix for failing test.

* Vertical level changes for profile.

Recent changes to behavior of specifying levels did not make it to the
profileData class. This should fix it, and unify those changes to avoid similar
disruptions moving forward.

- Added get_level for UPPData to unify the methods of getting a vertical level
for a given variable.
- Move the lev_val adjustment to the numeric_level method to reduce need for
repeats.
- vector_magnitude needs kwargs argument to allow calls with multiple
transforms.

* Clean up from recent changes/merges.

* Remove contour labels for certain contours.

This is similar logic to that in the title section. Don't label the contours for
hatched fields or for highlighting specific levels of the filled contour field.

Also adding a try statement in case no matching contours are found -- somewhat
innocuous -- so continue on with creating the map.

* Clean up for vertical levs.

* Minor behavior changes for driver script.

* Changes to image lists to reflect new names/levels.

* Moving a couple files.

* Updating tests, and linting.

* Reorder some lists.

* Remove unused fields in image lists.