Changes that allow reading in an ensemble.
These changes allow a user to specify that they'd like to run an ensemble panel
plot and then provide a templated path to the ensemble memebers. It'll load
those ensmeble members into an xarray via the GribFiles object. It stops at
printing the contents of the xarray.

Remaining work includes using that xarray to build a 3x4 panel plot to display
the data.