Add ensemble panel plots to support RRFSE. (#157)
* Changes that allow reading in an ensemble.

These changes allow a user to specify that they'd like to run an ensemble panel
plot and then provide a templated path to the ensemble memebers. It'll load
those ensmeble members into an xarray via the GribFiles object. It stops at
printing the contents of the xarray.

Remaining work includes using that xarray to build a 3x4 panel plot to display
the data.

* 1st raft of multiplanel plots

* add comment

* add sample notebook

* change from default_specs:level to cla.multipanel

* WIP on plotting panels.

Fails on needing 2D array instead of 3D.

* Adding a helper script for debugging.

* Will create panels with data in the right places.

* Adding small image list.

* Clean up obs panel addition

* Mostly linted.

* changes to ens plots - logo, colorbar, title

* Full image list to match 2021 RRFSE NCL graphics.

* Convert PNG to remove libpng errors.

* A bit of cleanup

* Refactor the maps/multipanel maps.

* Linting. Needed some re-arranging this time.

* Removing some utility scripts.

* Minor run-time fixes.

* Fixing copy/paste blunder.

* Fix for running single panel maps

* Update the levels to include "obs"

* Reduce degrees of freedom in conda env

* Linting.

Co-authored-by: Craig Hartsough <craig.hartsough@noaa.gov>